[components] Separate pure schema and resolution engine into two different submodules
