+++ python_modules/libraries/dagster-components/dagster_components/__init__.py
@@ -13,15 +13,17 @@ from dagster_components.core.component_scaffolder import (
     ComponentScaffoldRequest as ComponentScaffoldRequest,
     DefaultComponentScaffolder as DefaultComponentScaffolder,
 )
+from dagster_components.core.resolution_engine.context import ResolutionContext as ResolutionContext
+from dagster_components.core.resolution_engine.resolver import (
+    Resolver as Resolver,
+    resolver as resolver,
+)
 from dagster_components.core.schema.base import ComponentSchema as ComponentSchema
-from dagster_components.core.schema.context import ResolutionContext as ResolutionContext
 from dagster_components.core.schema.metadata import SchemaFieldInfo as SchemaFieldInfo
 from dagster_components.core.schema.objects import (
     AssetAttributesSchema as AssetAttributesSchema,
     AssetSpecSchema as AssetSpecSchema,
     AssetSpecTransformSchema as AssetSpecTransformSchema,
     OpSpecSchema as OpSpecSchema,
-    Resolver as Resolver,
-    resolver as resolver,
 )
 from dagster_components.version import __version__ as __version__
