+++ python_modules/libraries/dagster-components/dagster_components/utils.py
@@ -15,7 +15,7 @@ from dagster._core.definitions.declarative_automation.automation_condition impor
 )
 from dagster._core.errors import DagsterError
 
-from dagster_components.core.schema.context import ResolutionContext
+from dagster_components.core.resolution_engine.context import ResolutionContext
 from dagster_components.core.schema.objects import AssetAttributesSchema
 
 CLI_BUILTIN_COMPONENT_LIB_KEY = "builtin_component_lib"
